refactor: implement lazy import system and package reorganization
- Convert package __init__.py files to use lazy imports with __getattr__
- Reorganize module imports to prevent circular dependencies
- Update package configuration in pyproject.toml and setup.py
- Regenerate egg-info metadata for updated package structure
- Improve pytest configuration for better test discovery
- Enhance visualization module with better organization

This refactoring improves package loading performance and resolves
import issues during test discovery while maintaining full API compatibility.

+++ ergodic_insurance/src/ergodic_analyzer.py
@@ -17,7 +17,6 @@ from scipy import stats
 from .simulation import SimulationResults
 
 if TYPE_CHECKING:
-    from .insurance import InsuranceStructure
     from .insurance_program import InsuranceProgram
     from .loss_distributions import LossData
 
@@ -571,8 +570,8 @@ class ErgodicAnalyzer:
             }
 
         # Calculate ergodic divergence
-        time_avg_mean = np.mean(valid_time_avg) if valid_time_avg else -np.inf
-        ensemble_mean = ensemble_stats["mean"]
+        time_avg_mean = float(np.mean(valid_time_avg)) if valid_time_avg else -np.inf
+        ensemble_mean = float(ensemble_stats["mean"])
         ergodic_divergence = time_avg_mean - ensemble_mean
 
         # Validate results
